Branch Docs: Added: .standardese-config .hyde-config (Work in progress)
Documentation is scetchy as at best.
Standardese works but more work has to be done both in the comments and in configuration.
Type aliases aren't parse. And the header guard is invasive to the  docs.
Hyde doesn't wnat to work at the moment due to clang not understanding ranges entirely.